Revert "channelz: stage 4 - add security and socket option info (#2098)"
This reverts commit c1a21e22a684edc80bb02213ad1bb45c25bacbe3.
