+++ channelz/types.go
@@ -23,7 +23,6 @@ import (
 	"time"
 
 	"google.golang.org/grpc/connectivity"
-	"google.golang.org/grpc/credentials"
 	"google.golang.org/grpc/grpclog"
 )
 
@@ -282,9 +281,9 @@ type SocketInternalMetric struct {
 	RemoteAddr net.Addr
 	// Optional, represents the name of the remote endpoint, if different than
 	// the original target name.
-	RemoteName    string
-	SocketOptions *SocketOptionData
-	Security      credentials.ChannelzSecurityValue
+	RemoteName string
+	//TODO: socket options
+	//TODO: Security
 }
 
 // Socket is the interface that should be satisfied in order to be tracked by
